vuln_scanner/requirements.txt
python-nmap
scapy
reportlab
requests
colorama
